Move methods from the hermes helper module to either the hermes adapter or to common utils
